toolkit/components/passwordmgr/test/mochitest/test_autocomplete_autofill_related_realms_no_dupes.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test no duplicate logins using autofill/autocomplete with related realms</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
Login Manager test: no duplicate logins when using autofill and autocomplete with related realms
<script>
  addLoginsInParent(
    // Simple related domain relationship where example.com and other-example.com are in the related domains list
    ["https://other-example.com", "https://other-example.com", null, "relatedUser1", "relatedPass1", "uname", "pword"],

    // Example.com and example.co.uk are related, so sub.example.co.uk should appear on example.com's autocomplete dropdown
    // The intent is to cover the ebay.com/ebay.co.uk and all other country TLD cases
    // where the sign in page is actually signin.ebay.com/signin.ebay.co.uk but credentials could have manually been entered
    // for ebay.com/ebay.co.uk or automatically stored as signin.ebay.com/sigin.ebay.co.uk
    ["https://sub.example.co.uk", "https://sub.example.co.uk", null, "subUser1", "subPass1", "uname", "pword"],

    // Ensures there are no duplicates for the exact domain that the user is on
    ["https://example.com", "https://example.com", null, "exactUser1", "exactPass1", "uname", "pword"],
    ["https://www.example.com", "https://www.example.com", null, "exactWWWUser1", "exactWWWPass1", "uname", "pword"],
    );
</script>
<p id="display"></p>
<div id="content">
  <form id="form1" action="https://www.example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword">
    <button type="submit">Submit</button>
  </form>

</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Login Manager: related realms autofill/autocomplete. **/

const uname = getFormElementByName(1, "uname");
const pword = getFormElementByName(1, "pword");

function restoreForm() {
  uname.value = "";
  pword.value = "";
  uname.focus();
}

function sendFakeAutocompleteEvent(element) {
  var acEvent = document.createEvent("HTMLEvents");
  acEvent.initEvent("DOMAutoComplete", true, false);
  element.dispatchEvent(acEvent);
}

function spinEventLoop() {
  return Promise.resolve();
}

async function promiseACPopupClosed() {
  return SimpleTest.promiseWaitForCondition(async () => {
    let popupState = await getPopupState();
    return !popupState.open;
  }, "Wait for AC popup to be closed");
}

add_setup(async () => {
  listenForUnexpectedPopupShown();
});

add_task(async function test_no_duplicates_autocomplete_autofill() {
  await promiseFormsProcessedInSameProcess();
  await SimpleTest.promiseFocus(window);
  let shownPromise = promiseACShown();
  checkLoginForm(uname, "exactUser1", pword, "exactPass1")
  restoreForm();
  let results = await shownPromise;

  let popupState = await getPopupState();
  is(popupState.selectedIndex, -1, "Check no entires are selected upon opening");
  let expectedMenuItems = ["exactUser1", "exactWWWUser1", "relatedUser1", "subUser1"];
  checkAutoCompleteResults(results, expectedMenuItems, window.location.host, "Check all menuitems are displayed correctly");

  let acEvents = await getTelemetryEvents({ process: "parent", filterProps: TelemetryFilterPropsAC, clear: true });
  is(acEvents.length, 1, "One autocomplete event");
  checkACTelemetryEvent(acEvents[0], uname, {
    "hadPrevious": "0",
    "login": expectedMenuItems.length + "",
    "loginsFooter": "1"
  });
  restoreForm();
  synthesizeKey("KEY_Escape");
});
</script>
</pre>
</body>
</html>
